[disjoint][equals][strategies] Move CS-specific code of P/P disjoint/equals to strategies.
